refactor: move app view
- the idea is to keep related things together as for other UI projects
- app: defines entrypoint + contacts and DL modules entrypoint
- contact-groups: defines all contact-groups related views
- distribution-list: same as above but for dl
