+++ src/views/app/contacts-view.tsx
@@ -12,11 +12,11 @@ import { Redirect, Switch, Route, useRouteMatch } from 'react-router-dom';
 import { useUpdateView } from '../../carbonio-ui-commons/hooks/use-update-view';
 
 const LazyFolderView = lazy(
-	() => import(/* webpackChunkName: "folder-panel-view" */ './app/folder-view')
+	() => import(/* webpackChunkName: "folder-panel-view" */ '../../legacy/views/app/folder-view')
 );
 
 const LazyContactGroups = lazy(
-	() => import(/* webpackChunkName: "folder-panel-view" */ '../../views/contact-groups/cg-view')
+	() => import(/* webpackChunkName: "folder-panel-view" */ '../contact-groups/cg-view')
 );
 
 const ContactsView = (): React.JSX.Element => {
